Make pushed batch size configurable
In order to mimic real-world scenarios more realistically we want to
have the batch size of a push request configurable instead of the number
of log lines.

This change also adds two custom metrics which track the size of the
uncompressed log line data as well as the amount of lines pushed.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>