Reduce the amount of max active streams
The maximum amount of active streams is defined by the product of all
label values. This value was extremely high by default (6 × 3 × 5 × 10 × 50 × 177 × 997 = 7,941,105,000)

By making the cardinality of all label (except format and os) configurable, one can also control the
maximum amount of unique streams that can be produced.

Labels and their max amount of values are:

| name      | max values      |
| ========= | =============== |
| instance  | 1 per k6 worker | (not configurable)
| format    | 6               | (not configurable)
| os        | 3               | (not configurable)
| namespace | 169             |
| app       | ?               |
| pod       | 169             |
| language  | 177             |
| word      | 997             |

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>